# backend/src/main/java/com/jioha/asset/auth/AuthSignupRequest.java
package com.jioha.asset.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record AuthSignupRequest(
    @Email @NotBlank String email,
    @NotBlank @Size(min = 8) String password
) {
}