# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Asset Management API
  version: 0.2.0-mvp
  description: |
    MVP API contract aligned with PERSONAL_ASSET_PWA_GUIDE.md sections 8~14.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSignupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      parameters:
        - $ref: '#/components/parameters/CsrfTokenHeader'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    CsrfTokenHeader:
      in: header
      name: X-XSRF-TOKEN
      required: false
      schema:
        type: string
      description: |
        CSRF token header for session-cookie auth.
        Server sets cookie `XSRF-TOKEN` and the client should echo it in this header for mutating requests.
        MVP (same-origin + session cookie) expects this header for POST/PATCH/DELETE.
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
      example:
        error:
          code: VALIDATION_ERROR
          message: Invalid request.
          fieldErrors:
            - field: email
              reason: must be a well-formed email address
    FieldError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
...
    AuthSignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      example:
        email: demo@example.com
        password: demo-password
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      example:
        email: demo@example.com
        password: demo
    AuthMeResponse:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
      example:
        id: 1
        email: demo@example.com
...